+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# svgparser 0.1.2  2021-12-25
+
+* `viewBox` parsing should accept "," as number delimiter
+
 # svgparser 0.1.1  2021-12-05
 
 * Added initial support for `a`, `switch` and `image` tags